+++ crates/metrics/Cargo.toml
@@ -31,6 +31,7 @@ csv = "=1.1"
 structopt = "=0.3.3"
 strip-ansi-escapes = "=0.1.0"
 snowflake = "=1.3.0"
+crossbeam-channel = "=0.3.8"
 
 [dev-dependencies]
 tempfile = "=3.0.7"
