+++ crates/metrics/src/cloudwatch.rs
@@ -12,7 +12,7 @@ use std::{
     time::{SystemTime, UNIX_EPOCH},
 };
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use structopt::StructOpt;
 
 use rusoto_sts::{StsAssumeRoleSessionCredentialsProvider, StsClient};
@@ -78,6 +78,30 @@ pub struct QueryArgs {
         about = "The log stream pattern to filter messages over"
     )]
     pub log_stream_pat: Option<String>,
+}
+
+#[derive(Clone, Debug, Default, StructOpt)]
+pub struct CloudwatchLogsOptions {
+        #[structopt(
+            name = "region",
+            short = "r",
+            help = "The AWS region, defaults to eu-central-1."
+        )]
+        pub region: Option<Region>,
+        #[structopt(
+            name = "log_group_name",
+            short = "l",
+            help = "The AWS log group name to query over."
+        )]
+        pub log_group_name: Option<String>,
+        #[structopt(
+            name = "assume_role_arn",
+            short = "a",
+            help = "Optional override for the amazon role to assume when querying"
+        )]
+        pub assume_role_arn: Option<String>,
+        #[structopt(flatten)]
+        pub query_args: QueryArgs,
 }
 
 impl CloudWatchLogger {
@@ -312,7 +336,7 @@ impl CloudWatchLogger {
     }
 }
 
-const PUBLISH_CHUNK_SIZE: usize = 100;
+const PUBLISH_CHUNK_SIZE: usize = 10;
 
 impl MetricPublisher for CloudWatchLogger {
     fn publish(&mut self, metric: &Metric) {
@@ -414,86 +438,6 @@ impl Default for CloudWatchLogger {
             &DEFAULT_REGION,
         )
     }
-}
-
-const LOG_STREAM_SEPARATOR: &str = ".";
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct ScenarioData {
-    run_name: String,
-    net_name: String,
-    dna_name: String,
-    scenario_name: String,
-    player_name: String,
-}
-
-impl Into<String> for ScenarioData {
-    fn into(self: Self) -> String {
-        let s = self;
-        format!(
-            "{}.{}.{}.{}.{}",
-            s.run_name, s.net_name, s.dna_name, s.scenario_name, s.player_name
-        )
-    }
-}
-
-impl TryFrom<String> for ScenarioData {
-    type Error = String;
-
-    fn try_from(s: String) -> Result<Self, Self::Error> {
-        let split = s.split(LOG_STREAM_SEPARATOR).collect::<Vec<_>>();
-        if split.len() < 4 {
-            return Err(format!(
-                "Log stream name doesn't have at least 4 path elements: {:?}",
-                split
-            ));
-        }
-        Ok(Self {
-            run_name: split[0].into(),
-            net_name: split[1].into(),
-            dna_name: split[2].into(),
-            scenario_name: split[3].into(),
-            player_name: split[4].into(),
-        })
-    }
-}
-
-impl TryFrom<LogStream> for ScenarioData {
-    type Error = String;
-    fn try_from(log_stream: LogStream) -> Result<Self, Self::Error> {
-        let result: Result<Self, Self::Error> = log_stream
-            .log_stream_name
-            .map(|x| Ok(x))
-            .unwrap_or_else(|| Err("Log stream name missing".into()))
-            .and_then(TryFrom::try_from);
-        result
-    }
-}
-impl ScenarioData {
-    /// Groups by everything _but_ the player name
-    fn without_player_name(&self) -> String {
-        format!(
-            "{}.{}.{}.{}",
-            self.run_name, self.net_name, self.dna_name, self.scenario_name
-        )
-    }
-}
-
-/// Groups a log stream name by its scenario name, which is by convention is the 2nd to last field.
-/// Eg. "2019-12-06_01-54-47_stress_10_1_2.sim2h.smoke.9"
-pub fn group_by_scenario(
-    log_stream_names: &mut dyn Iterator<Item = String>,
-) -> HashMap<String, HashSet<ScenarioData>> {
-    log_stream_names.fold(HashMap::new(), |mut grouped, log_stream_name| {
-        let scenario_data: Result<ScenarioData, _> = log_stream_name.try_into();
-        if let Ok(scenario_data) = scenario_data {
-            grouped
-                .entry(scenario_data.without_player_name())
-                .or_insert_with(HashSet::new)
-                .insert(scenario_data);
-        }
-        grouped
-    })
 }
 
 pub const FINAL_EXAM_NODE_ROLE: &str =
