+++ crates/metrics/src/metrics.rs
@@ -2,19 +2,21 @@ use holochain_locksmith::RwLock;
 /// Metric suppport for holochain. Provides metric representations to
 /// sample, publish, aggregate, and analyze metric data.
 use std::sync::Arc;
+use crossbeam_channel::*;
 
 /// Represents a single sample of a numerical metric determined by `name`.
-// TODO Consider renaming to Sample
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Metric {
-    pub name: String,
-    pub value: f64,
+   pub name: String,
+   pub stream_id: Option<String>,
+   pub value: f64,
 }
 
 impl Metric {
-    pub fn new(name: &str, value: f64) -> Self {
+    pub fn new<S:Into<String>, S2:Into<Option<String>>>(name: S, stream_id: S2, value: f64) -> Self {
         Self {
-            name: name.to_string(),
+            name: name.into(),
+            stream_id: stream_id.into(),
             value,
         }
     }
@@ -24,6 +26,20 @@ impl Metric {
 pub trait MetricPublisher: Sync + Send {
     /// Publish a single metric.
     fn publish(&mut self, metric: &Metric);
+}
+
+
+pub struct QueuedPublisher<M:MetricPublisher> {
+    metric_publisher: M,
+    sender: Sender<Metric>,
+    receiver: Receiver<Metric>
+}
+
+impl<M:MetricPublisher> MetricPublisher for QueuedPublisher<M> {
+
+    fn publish(&mut self, metric:&Metric) {
+       self.sender.send(*metric).unwrap();
+    }
 }
 
 /// The default metric publisher trait implementation
@@ -42,7 +58,8 @@ macro_rules! with_latency_publishing {
 
         let metric_name = format!("{}.latency", $metric_prefix);
 
-        let metric = $crate::Metric::new(metric_name.as_str(), latency as f64);
+        // TODO pass in stream id or not?
+        let metric = $crate::Metric::new(metric_name.as_str(), None, latency as f64);
         $publisher.write().unwrap().publish(&metric);
         ret
     }}
@@ -89,7 +106,7 @@ mod test {
     #[test]
     fn can_publish_to_logger() {
         let mut publisher = crate::logger::LoggerMetricPublisher;
-        let metric = Metric::new("latency", 100.0);
+        let metric = Metric::new("latency", None, 100.0);
 
         publisher.publish(&metric);
     }
